+++ data-raw/GenerateSysData.R
@@ -44,8 +44,8 @@ MK_RS5NF_Distance <- calcTargetingDistance(model=MK_RS5NF)
 e <- new.env()
 load("data-raw/HS1F_v0.1.4.rda", envir=e)
 load("data-raw/M1N_v0.1.4.rda", envir=e)
-HS1F_Distance_v0.1.4 <- get("HS1FDistance", envir=e)
-M1N_Distance_v0.1.4 <- get("M1NDistance", envir=e)
+HS1F_Compat <- get("HS1FDistance", envir=e)
+M1N_Compat <- get("M1NDistance", envir=e)
 rm(e)
 
 #### Load other saved data ####
@@ -71,6 +71,6 @@ devtools::use_data(NUCLEOTIDES,
                    HKL_S5F_Distance,
                    MK_RS1NF_Distance,
                    MK_RS5NF_Distance,
-                   HS1F_v0.1.4,
-                   M1N_v0.1.4,
+                   HS1F_Compat,
+                   M1N_Compat,
                    internal=TRUE, overwrite=TRUE)