+++ tests/testthat/test.distToNearest.R
@@ -5,7 +5,7 @@ load(file.path("..", "data-tests", "TestDb.rda"), envir=e1)
 db <- get("TestDb", envir=e1)
 rm(e1)
 
-test_that("Test cross distToNearest with model hs1f", {
+test_that("Test cross distToNearest with model hh_s1f", {
     ## Reproduce example
     db <- subset(db, CPRIMER %in% c("IGHA","IGHM") & 
                      BARCODE %in% c("RL016","RL018","RL019","RL021"))
@@ -17,10 +17,10 @@ test_that("Test cross distToNearest with model hs1f", {
     
     test_idx <- c(1:5,10:15,300:305)
     
-    ## Test hs1f 
+    ## Test hh_s1f 
     
     dist_hs1f <- distToNearest(db, vCallColumn="V_CALL_GENOTYPED", 
-                                    model="hs1f", first=FALSE, normalize="length")
+                                    model="hh_s1f", first=FALSE, normalize="length")
     ## Test if the updated function reproduces results
     expect_equal(dist_hs1f$DIST_NEAREST[test_idx],
                  c(NA,NA,NA,NA,0.4040, 0.4447, 0.3963, 0.3469, 0.3050, 0.3050,
@@ -29,18 +29,18 @@ test_that("Test cross distToNearest with model hs1f", {
     
     ## There's only one donor, hence cross-donor will return all NA
     dist_hs1f_cross_donor <- distToNearest(db2, vCallColumn="V_CALL_GENOTYPED", 
-                                           model="hs1f", first=FALSE, normalize="none", nproc=1, cross="DONOR")
+                                           model="hh_s1f", first=FALSE, normalize="none", nproc=1, cross="DONOR")
 
     expect_true(all(is.na(dist_hs1f_cross_donor$CROSS_DIST_NEAREST)))
     
     ## fields=NULL and fields=DONOR should give same results
     cross_dist_hs1f <- distToNearest(db2, vCallColumn="V_CALL_GENOTYPED", 
-                                     model="hs1f", first=FALSE, normalize="length",nproc=1,
+                                     model="hh_s1f", first=FALSE, normalize="length",nproc=1,
                                      cross="SAMPLE")
     
     
     cross_dist_hs1f_donor <- distToNearest(db2, vCallColumn="V_CALL_GENOTYPED", 
-                               model="hs1f", first=FALSE, normalize="length",nproc=1,
+                               model="hh_s1f", first=FALSE, normalize="length",nproc=1,
                                cross="SAMPLE", fields="DONOR")
     
     expect_equal(cross_dist_hs1f, cross_dist_hs1f_donor)
@@ -63,17 +63,17 @@ test_that("Test cross distToNearest with model hs1f", {
     # t(db3[c(1,316,nrow(db3)),c("JUNCTION","V_CALL","J_CALL","DONOR","SAMPLE")] )
     
     db2_1_316_630 <- distToNearest(db2[c(1,316,630),], vCallColumn="V_CALL_GENOTYPED", 
-                  model="hs1f", first=FALSE, normalize="length",cross="SAMPLE")
+                  model="hh_s1f", first=FALSE, normalize="length",cross="SAMPLE")
     ## Exactly same seq, returns NA
     expect_equal(db2_1_316_630$CROSS_DIST_NEAREST,c(NA,NA,NA))
     
     ## One seq has been edited, will return distance values
     db3_1_316_630 <- distToNearest(db3[c(1,316,630),], vCallColumn="V_CALL_GENOTYPED", 
-                             model="hs1f", first=FALSE, normalize="length",cross="SAMPLE")
+                             model="hh_s1f", first=FALSE, normalize="length",cross="SAMPLE")
     expect_equal(db3_1_316_630$CROSS_DIST_NEAREST,c(0.0175,0.0175,0.0175), tolerance=0.001)
 })
 
-test_that("Test cross distToNearest with model hs5f", {
+test_that("Test cross distToNearest with model hh_s5f", {
     ## Reproduce vignette
     db <- subset(db, CPRIMER %in% c("IGHA","IGHM") & 
                      BARCODE %in% c("RL016","RL018","RL019","RL021"))
@@ -86,10 +86,10 @@ test_that("Test cross distToNearest with model hs5f", {
     
     test_idx <- c(1:5,10:15,300:305)
     
-    ## Test hs1f 
+    ## Test hh_s5f 
     
     dist_hs5f <- distToNearest(db, vCallColumn="V_CALL_GENOTYPED", 
-                               model="hs5f", first=FALSE, normalize="none", nproc=1)
+                               model="hh_s5f", first=FALSE, normalize="none", nproc=1)
 
     ## Test if the updated function reproduces results
     expect_equal(dist_hs5f$DIST_NEAREST[test_idx],
@@ -100,15 +100,15 @@ test_that("Test cross distToNearest with model hs5f", {
     
     ## There's only one donor, hence cross-donor will return all NA
     dist_hs5f_cross_donor <- distToNearest(db2, vCallColumn="V_CALL_GENOTYPED", 
-                               model="hs5f", first=FALSE, normalize="none", nproc=1, cross="DONOR")
+                               model="hh_s5f", first=FALSE, normalize="none", nproc=1, cross="DONOR")
     expect_true(all(is.na(dist_hs5f_cross_donor$CROSS_DIST_NEAREST)))
     
     ## fields=NULL and fields=DONOR should give same results
     cross_dist_hs5f <- distToNearest(db2, vCallColumn="V_CALL_GENOTYPED", 
-                                     model="hs5f", first=FALSE, normalize="none",nproc=1,
+                                     model="hh_s5f", first=FALSE, normalize="none",nproc=1,
                                      cross="SAMPLE")
     cross_dist_hs5f_donor <- distToNearest(db2, vCallColumn="V_CALL_GENOTYPED", 
-                                           model="hs5f", first=FALSE, normalize="none",nproc=1,
+                                           model="hh_s5f", first=FALSE, normalize="none",nproc=1,
                                            cross="SAMPLE", fields="DONOR")
     expect_equal(cross_dist_hs5f, cross_dist_hs5f_donor)
     
@@ -132,13 +132,13 @@ test_that("Test cross distToNearest with model hs5f", {
     # t(db3[c(1,316,nrow(db3)),c("JUNCTION","V_CALL","J_CALL","DONOR","SAMPLE")] )
     
     db2_1_316_630_hs5f <- distToNearest(db2[c(1,316,630),], vCallColumn="V_CALL_GENOTYPED", 
-                                   model="hs5f", first=FALSE, normalize="none",cross="SAMPLE")
+                                   model="hh_s5f", first=FALSE, normalize="none",cross="SAMPLE")
     ## Exactly same seq, returns NA
     expect_equal(db2_1_316_630_hs5f$CROSS_DIST_NEAREST,c(NA,NA,NA))
     
     ## One seq has been edited, will return distance values
     db3_1_316_630_hs5f <- distToNearest(db3[c(1,316,630),], vCallColumn="V_CALL_GENOTYPED", 
-                                   model="hs5f", first=FALSE, normalize="none",cross="SAMPLE")
+                                   model="hh_s5f", first=FALSE, normalize="none",cross="SAMPLE")
     expect_equal(db3_1_316_630_hs5f$CROSS_DIST_NEAREST,c(1.001,1.001,1.001), tolerance=0.001)
     
     seq1 <- c("NNACG", "NACGT", "ACGTA", "CGTAC", "GTACG", "TACGT", "ACGTA", 
@@ -164,7 +164,7 @@ test_that("Test distToNearest with unrecognized characters", {
     ## This seq belongs to clone of size 1
     db2$JUNCTION[3] <- gsub("T","Z",db2$JUNCTION[5] )
     expect_warning(
-        dist_hs1f <- distToNearest(db2,model="hs1f", first=FALSE, normalize="length"),
+        dist_hs1f <- distToNearest(db2, model="hh_s1f", first=FALSE, normalize="length"),
         "Invalid sequence characters"
     )
     
@@ -172,7 +172,7 @@ test_that("Test distToNearest with unrecognized characters", {
     ## This seq belongs to clone of size > 1
     db2$JUNCTION[5] <- gsub("T","Z",db$JUNCTION[5] )
     expect_warning(
-        dist_hs1f <- distToNearest(db2,model="hs1f", first=FALSE, normalize="length"),
+        dist_hs1f <- distToNearest(db2, model="hh_s1f", first=FALSE, normalize="length"),
         "Invalid sequence characters"
     )